+++ app/backends/jwt.py
@@ -1,0 +1,46 @@
+import jwt
+
+from django.conf import settings
+
+from rest_framework import authentication, exceptions
+
+from app.models import User
+
+
+class JWTAuthentication(authentication.BaseAuthentication):
+    authentication_header_prefix = "Token"
+
+    def authenticate(self, request):
+        request.user = None
+
+        auth_header = authentication.get_authorization_header(request).split()
+        auth_header_prefix = self.authentication_header_prefix.lower()
+
+        if not auth_header:
+            return None
+
+        if len(auth_header) == 1:
+            return None
+
+        elif len(auth_header) > 2:
+            return None
+
+        prefix = auth_header[0].decode('utf-8')
+        token = auth_header[1].decode('utf-8')
+
+        if prefix.lower() != auth_header_prefix:
+            return None
+
+        try:
+            payload = jwt.decode(token, settings.SECRET_KEY)
+        except Exception:
+            msg = 'Token decode error'
+            raise exceptions.AuthenticationFailed(msg)
+
+        try:
+            user = User.objects.get(pk=payload['id'])
+        except User.DoesNotExist:
+            msg = 'No user for this token'
+            raise exceptions.AuthenticationFailed(msg)
+
+        return user, token
