+++ app/models/__init__.py
@@ -1,0 +1,28 @@
+import jwt
+from django.contrib.auth.base_user import AbstractBaseUser
+from django.contrib.auth.models import PermissionsMixin
+from django.db import models
+
+from app.managers.user import UserManager
+from core import settings
+
+
+class User(AbstractBaseUser, PermissionsMixin):
+    username = models.CharField(db_index=True, max_length=255, unique=True)
+    email = models.EmailField(db_index=True, unique=True)
+
+    USERNAME_FIELD = "email"
+    REQUIRED_FIELDS = ["username"]
+
+    objects = UserManager()
+
+    @property
+    def token(self):
+        return self._generate_jwt_token()
+
+    def _generate_jwt_token(self):
+        token = jwt.encode({
+            "id": self.pk,
+        }, settings.SECRET_KEY, algorithm="HS256")
+
+        return token.decode("utf-8")
