+++ app/serializers/login.py
@@ -1,0 +1,23 @@
+from django.contrib.auth import authenticate
+from rest_framework import serializers
+
+
+class LoginSerializer(serializers.Serializer):
+    email = serializers.CharField(max_length=255)
+    password = serializers.CharField(max_length=128, write_only=True)
+    access_token = serializers.CharField(max_length=255, read_only=True)
+
+    def validate(self, data):
+        email = data.get('email')
+        password = data.get('password')
+
+        user = authenticate(username=email, password=password)
+
+        if user is None:
+            raise serializers.ValidationError('A user with this email and password was not found')
+
+        return {
+            'email': user.email,
+            'username': user.username,
+            'access_token': user.token
+        }
