+++ app/serializers/registration.py
@@ -1,0 +1,18 @@
+from rest_framework import serializers
+
+from app.models import User
+
+
+class RegistrationSerializer(serializers.ModelSerializer):
+    password = serializers.CharField(
+        max_length=128,
+        min_length=8,
+        write_only=True
+    )
+
+    class Meta:
+        model = User
+        fields = ['email', 'username', 'password']
+
+    def create(self, validated_data):
+        return User.objects.create_user(**validated_data)
