+++ app/urls.py
@@ -1,1 +1,9 @@
-urlpatterns = []
+from django.urls import path
+
+from app.views.login import LoginView
+from app.views.registration import RegistrationView
+
+urlpatterns = [
+    path("register", RegistrationView.as_view()),
+    path("login", LoginView.as_view())
+]
