+++ core/asgi.py
@@ -1,18 +1,21 @@
 import os
 
-from channels.auth import AuthMiddlewareStack
+from channels.security.websocket import AllowedHostsOriginValidator
 from django.core.asgi import get_asgi_application
 from channels.routing import ProtocolTypeRouter, URLRouter
 
 import app.routing
+from core.middleware.channels import JwtAuthMiddlewareStack
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "core.settings")
 
 application = ProtocolTypeRouter({
     "http": get_asgi_application(),
-    "websocket": AuthMiddlewareStack(
-        URLRouter(
-            app.routing.websocket_urlpatterns
-        )
+    "websocket": AllowedHostsOriginValidator(
+        JwtAuthMiddlewareStack(
+            URLRouter(
+                app.routing.websocket_urlpatterns
+            )
+        ),
     )
 })
