+++ core/middleware/channels.py
@@ -1,0 +1,45 @@
+from urllib.parse import parse_qs
+
+from channels.auth import AuthMiddlewareStack
+from channels.db import database_sync_to_async
+from channels.middleware import BaseMiddleware
+from django.contrib.auth import get_user_model
+from django.contrib.auth.models import AnonymousUser
+from django.db import close_old_connections
+from jwt import decode
+from rest_framework_simplejwt.exceptions import InvalidToken, TokenError
+from rest_framework_simplejwt.tokens import UntypedToken
+
+from app.models import User
+from core import settings
+
+
+@database_sync_to_async
+def get_user(token):
+    try:
+        return get_user_model().objects.get(id=token["user_id"])
+    except User.DoesNotExist:
+        return AnonymousUser()
+
+
+class JwtAuthMiddleware(BaseMiddleware):
+
+    async def __call__(self, scope, receive, send):
+        close_old_connections()
+
+        token = parse_qs(scope["query_string"].decode("utf8"))["token"][0]
+
+        try:
+            UntypedToken(token)
+        except (InvalidToken, TokenError) as e:
+            return None
+        else:
+            decoded_data = decode(token, settings.SECRET_KEY, algorithms=["HS256"])
+
+        scope['user'] = await get_user(validated_token=decoded_data)
+
+        return await super.__call__(scope, receive, send)
+
+
+def JwtAuthMiddlewareStack(inner):
+    return JwtAuthMiddleware(AuthMiddlewareStack(inner))
